refactor(sync): isolate adoption snapshot policy
