+++ config/architecture-budgets.json
@@ -7,7 +7,6 @@
     "crates/connect-agent/src/mirrors.rs": 1594,
     "crates/connect-agent/src/server.rs": 1967,
     "crates/connect-mirror/src/lib.rs": 1915,
-    "packages/sync/src/adoption.ts": 1127,
     "services/server/src/legacy-baseline.ts": 1114
   }
 }
